Move interface of plot into Solution
