+++ examples/getting_started/custom_phase_transitions.py
@@ -22,7 +22,6 @@ from bioptim import (
     InitialGuessList,
     PhaseTransitionFcn,
     PhaseTransitionList,
-    ShowResult,
     OdeSolver,
 )
 
@@ -176,5 +175,4 @@ if __name__ == "__main__":
     sol = ocp.solve(show_online_optim=True)
 
     # --- Show results --- #
-    result = ShowResult(ocp, sol)
-    result.animate()
+    sol.animate()
