+++ examples/getting_started/example_cyclic_movement.py
@@ -22,7 +22,6 @@ from bioptim import (
     Bounds,
     QAndQDotBounds,
     InitialGuess,
-    ShowResult,
     OdeSolver,
     PhaseTransitionList,
     PhaseTransitionFcn,
@@ -123,5 +122,4 @@ if __name__ == "__main__":
     sol = ocp.solve(show_online_optim=True)
 
     # --- Show results --- #
-    result = ShowResult(ocp, sol)
-    result.animate()
+    sol.animate()
