+++ examples/muscle_driven_ocp/static_arm_with_contact.py
@@ -17,7 +17,6 @@ from bioptim import (
     BoundsList,
     QAndQDotBounds,
     InitialGuessList,
-    ShowResult,
     OdeSolver,
 )
 
@@ -112,6 +111,5 @@ if __name__ == "__main__":
     sol = ocp.solve(show_online_optim=True)
 
     # --- Show results --- #
-    result = ShowResult(ocp, sol)
-    result.animate(show_meshes=False)
-    result.graphs()
+    sol.animate(show_meshes=False)
+    sol.graphs()
