+++ examples/optimal_time_ocp/pendulum_min_time_Mayer.py
@@ -18,7 +18,6 @@ from bioptim import (
     BoundsList,
     QAndQDotBounds,
     InitialGuessList,
-    ShowResult,
     OdeSolver,
 )
 
@@ -120,5 +119,4 @@ if __name__ == "__main__":
     # --- Show results --- #
     print(f"The optimized phase time is: {sol.parameters['time'][0, 0]}, good job Mayer!")
 
-    result = ShowResult(ocp, sol)
-    result.animate()
+    sol.animate()
