+++ examples/torque_driven_ocp/maximize_predicted_height_CoM.py
@@ -17,7 +17,6 @@ from bioptim import (
     BoundsList,
     QAndQDotBounds,
     InitialGuessList,
-    ShowResult,
     OdeSolver,
     Axis,
     ConstraintList,
@@ -161,5 +160,4 @@ if __name__ == "__main__":
     sol = ocp.solve(show_online_optim=True)
 
     # --- Show results --- #
-    result = ShowResult(ocp, sol)
-    result.animate(n_frames=40)
+    sol.animate(n_frames=40)
