+++ examples/torque_driven_ocp/track_markers_2D_pendulum.py
@@ -20,7 +20,6 @@ from bioptim import (
     BoundsList,
     QAndQDotBounds,
     InitialGuessList,
-    ShowResult,
     ObjectiveList,
     ObjectiveFcn,
     Axis,
@@ -211,5 +210,4 @@ if __name__ == "__main__":
     sol = ocp.solve(show_online_optim=True)
 
     # --- Show results --- #
-    result = ShowResult(ocp, sol)
-    result.animate()
+    sol.animate()
