+++ examples/torque_driven_ocp/trampo_quaternions.py
@@ -16,7 +16,6 @@ from bioptim import (
     BoundsList,
     QAndQDotBounds,
     InitialGuessList,
-    ShowResult,
     InterpolationType,
     OdeSolver,
 )
@@ -145,5 +144,4 @@ if __name__ == "__main__":
     print("\n")
 
     # Print the last solution
-    result_plot = ShowResult(ocp, sol)
-    result_plot.animate()
+    sol.animate()
